Breaking: De-archangel JophielConfiguration; fold into JudgelsServerConfiguration
