+++ deployment/v3-ubuntu-24.04/ansible/conf/judgels-server.yml.j2
@@ -64,7 +64,6 @@ judgels:
     gradingRequestQueueName: judgels-grading-request
     gradingResponseQueueName: judgels-grading-response
 
-jophiel:
   auth:
 {% if jophiel_auth_google_clientId is defined %}
     google:
