+++ judgels-backends/judgels-server-app/var/conf/judgels-server.yml.example
@@ -58,7 +58,6 @@ judgels:
     gradingRequestQueueName: judgels-grading-request
     gradingResponseQueueName: judgels-grading-response
 
-jophiel:
   userResetPassword:
     enabled: false
     requestEmailTemplate:
